refactor: rewire source imports from facade to taxonomy modules
All 5 source files now import directly from taxonomy-store, taxonomy-selectors, or taxonomy-mutations. category-manager.js is no longer imported by any production code.
